+++ mike/driver.py
@@ -112,7 +112,8 @@ def list_versions(args):
 
 def set_default(args):
     check_remote_status(args, strict=True)
-    commands.set_default(args.version, args.branch, args.message)
+    commands.set_default(args.version, args.template, args.branch,
+                         args.message)
     if args.push:
         git_utils.push_branch(args.remote, args.branch, args.force)
 
@@ -190,6 +191,8 @@ def main():
     )
     set_default_p.set_defaults(func=set_default)
     add_git_arguments(set_default_p)
+    set_default_p.add_argument('-t', '--template',
+                               help='the template file to use')
     set_default_p.add_argument('version', metavar='VERSION',
                                help='version to set as default')